==> data-raw/05-regular.Rmd <==
---
output: github_document
---

# `house_reps_regular`

* only includes *Regular* sessions,
* after OR became a state, with
* legislators in the house at the **start** of the session, and 
* their party affiliation at the **start** of the session.

```{r setup}
library(tidyverse)
library(here)
```

```{r}
house_reps <- here("data-raw", "chronology-patched.csv") %>% 
  read_csv(col_types = cols(desk_number = col_integer()))
```

Start when Oregon is a state, February 14, 1859.
```{r}
state_history <- house_reps %>% 
  filter(session_year >= 1859)
```

Use only regular sessions.
```{r}
regular <- state_history %>% 
  filter(regular)
```

## Split terms

`split-terms.csv` curated by hand contains legislators that share a session, e.g. one person leaves and is replaced by another:
```{r}
splits <- read_csv(here("data-raw", "split-terms.csv")) %>% 
  rename_all(tolower)
```

For example, the first recorded split term
```{r}
first_split <- splits %>% 
  left_join(regular) %>% 
  filter(split_id == min(split_id)) 

first_split %>% 
  select(session_year, legislator, split_order)

writeLines(first_split$notes)
```

Now remove second in seat:
```{r}
second_seat <- splits %>% 
  filter(split_order == 2) %>% 
  select(session_year, legislator)

regular_first_in_seat <- regular %>% 
  anti_join(second_seat) 

regular_first_in_seat %>% 
  group_by(session_year) %>% 
  count()
```

## Only use first party affiliation

```{r}
regular_no_mixed_party <- regular_first_in_seat %>% 
  mutate(
    party_mixed = party %>% str_detect("/"), 
    party = ifelse(party_mixed, map_chr(str_split(party, "/"), 1), party)
  ) %>% 
  select(-party_mixed)
```

## Output

```{r}
house_reps_regular <- regular_no_mixed_party
usethis::use_data(house_reps_regular, overwrite = TRUE)
```
